Crate tokenizerStable

Structs

Token
Tokens are representations of the items in a program.

Enums

TokenType
The different token types that are possible.

Functions

tokens
The tokens function is the center of the tokenizer. It splits, separates, checks, matadatas, and organizes each and every character in the code string.

ttype_of
Although small and fairly simple, ttype_of is the heart and soul of the tokenizer.