Enum tokenizer::TokenType

pub enum TokenType {
    String,
    Identifier,
    OpenParen,
    CloseParen,
    Lambda,
    Assign,
    Separator,
}

The different token types that are possible.

TokenType represents the different kinds of tokens there are. One of these is assigned to each of the tokens that are created.

Variants

String

Identifier

OpenParen

CloseParen

Lambda

Assign

Separator