Function tokenizer::tokensUnstable

pub fn tokens(string: &str) -> Vec<Token>

The tokens function is the center of the tokenizer. It splits, separates, checks, matadatas, and organizes each and every character in the code string.

Tokens goes through the given code, line by line, and spits each line up into characters, then, iterating over each character, examines it's type using ttype_of. Then, based on the type, and the action also associated with that token, takes action. Then, it pushes the token onto the token vector. There is a special tokenizer action assiciated with most special characters. These special token actions can only be perforemed within the tokenizer, otherwise it would be to late to make these expantions and changes.