Struct tokenizer::Token

pub struct Token {
    pub value: String,
    pub ttype: TokenType,
}

Tokens are representations of the items in a program.

Instead of parsing directly on the text, wich would lead to messy, mangled code, most languages use a tokenizer to seperate out concerns. The tokenizer figures out what a thing is, and builds more understandable ASTs out of that, assigning more metadata as it goes. Token is a struct that represents the basic metadata attached to a token.

Fields

value

ttype